Split unittests into separate files
