+++ test/GuzzleHttpMockTest/MockTest.php
@@ -2,33 +2,15 @@
 
 namespace Aeris\GuzzleHttpMock\Test\GuzzleHttpMockTest;
 
-use Aeris\GuzzleHttpMock\Exception\CompoundUnexpectedHttpRequestException;
-use Aeris\GuzzleHttpMock\Exception\UnexpectedHttpRequestException;
-use Exception;
-use GuzzleHttp\Client as GuzzleClient;
+use Psr\Http\Message\RequestInterface;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
-use Aeris\GuzzleHttpMock\Mock;
-use Aeris\GuzzleHttpMock\Expect;
 use GuzzleHttp\RequestOptions;
-use PHPUnit\Framework\TestCase;
-use Psr\Http\Message\RequestInterface;
+use Aeris\GuzzleHttpMock\Expect;
+use Aeris\GuzzleHttpMock\Exception\UnexpectedHttpRequestException;
 
-class MockTest extends TestCase
+class MockTest extends BaseTest
 {
-    /** @var GuzzleClient */
-    protected $guzzleClient;
-
-    /** @var Mock */
-    protected $httpMock;
-
-    protected function setUp(): void
-    {
-        $this->httpMock = new Mock();
-        $handlerStack = $this->httpMock->getHandlerStackWithMiddleware();
-        $this->guzzleClient = new GuzzleClient(['handler' => $handlerStack]);
-    }
-
     /** @test */
     public function shouldReturnAResponseForARequestObject()
     {
@@ -95,41 +77,6 @@ class MockTest extends TestCase
             ->put('http://example.com/foo', [
                 'query' => ['faz' => 'baz'],
                 'form_params' => ['shakeyo' => 'body'],
-            ]);
-
-        $this->httpMock->verify();
-        $this->assertSame($mockResponse, $actualResponse);
-    }
-
-    /** @test */
-    public function shouldReturnAResponseForAJsonBodyParamsExpectation()
-    {
-        $mockResponse = new Response(200,
-            ['Content-Type' => 'application/json'],
-            json_encode([
-                'hello' => 'world',
-                'howareyou' => 'today'
-            ])
-        );
-
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withUrl('http://example.com/foo')
-            ->withMethod('PUT')
-            ->withQueryParams([
-                'faz' => 'baz'
-            ])
-            ->withJsonBodyParams([
-                'shakeyo' => 'body'
-            ])
-            ->andRespondWith($mockResponse);
-
-
-        $actualResponse = $this->guzzleClient
-            ->put('http://example.com/foo', [
-                'query' => ['faz' => 'baz'],
-                'body' => json_encode(['shakeyo' => 'body']),
-                'headers' => ['Content-Type' => 'application/json']
             ]);
 
         $this->httpMock->verify();
@@ -356,377 +303,6 @@ class MockTest extends TestCase
     }
 
     /** @test */
-    public function shouldCompareQueryStringLiterally()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->withQueryString('param1=value1&param2=%20spaces')
-            ->andRespondWithJson(['bar' => 'foo']);
-
-        $response1 = $this->guzzleClient
-            ->get('http://www.example.com/foo?param1=value1&param2=%20spaces');
-
-        $this->assertEquals(['bar' => 'foo'], json_decode((string)$response1->getBody(), true));
-    }
-
-    /** @test */
-    public function url_notMatch_shouldFail()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo');
-
-        $this->expectException(UnexpectedHttpRequestException::class);
-        $this->guzzleClient
-            ->get('http://www.example.com/shazlooey');
-
-        $this->httpMock->verify();
-    }
-
-    /** @test */
-    public function url_customLogic_match_shouldPass()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl(function ($url) {
-                return preg_match('/foo$/', $url) === 1;
-            });
-
-        $this->guzzleClient
-            ->get('http://www.example.com/foo');
-
-        $this->assertTrue($this->httpMock->verify());
-    }
-
-    /** @test */
-    public function url_customLogic_notMatch_shouldFail()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl(function ($url) {
-                return preg_match('/foo$/', $url) === 1;
-            });
-
-
-        $this->guzzleClient
-            ->get('http://www.example.com/shablooey');
-
-        $this->expectException(UnexpectedHttpRequestException::class);
-        $this->httpMock->verify();
-    }
-
-    /** @test */
-    public function method_notMatch_shouldFail()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('POST')
-            ->withUrl('http://www.example.com/foo');
-
-        $this->expectException(UnexpectedHttpRequestException::class);
-        $this->guzzleClient
-            ->get('http://www.example.com/foo');
-
-        $this->httpMock->verify();
-    }
-
-    /** @test */
-    public function method_match_customLogic_shouldPass()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod(function ($method) {
-                return strlen($method) === 3;
-            })
-            ->withUrl('http://www.example.com/foo');
-
-        $this->guzzleClient
-            ->put('http://www.example.com/foo');
-
-        $this->assertTrue($this->httpMock->verify());
-    }
-
-    /** @test */
-    public function method_notMatch_customLogic_shouldFail()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod(function ($method) {
-                return strlen($method) === 3;
-            })
-            ->withUrl('http://www.example.com/foo');
-
-        $this->expectException(UnexpectedHttpRequestException::class);
-        $this->guzzleClient
-            ->post('http://www.example.com/foo');
-
-        $this->httpMock->verify();
-    }
-
-    /** @test */
-    public function query_notMatch_shouldFail()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->withQueryParams(['foo' => 'bar']);
-
-        $this->expectException(UnexpectedHttpRequestException::class);
-        $this->guzzleClient
-            ->get('http://www.example.com/shazlooey', [
-                'query' => ['not' => 'what I expected']
-            ]);
-
-        $this->httpMock->verify();
-    }
-
-    /** @test */
-    public function queryParams_customLogic_match_shouldPass()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->withQueryParams(function ($actualParams) {
-                return $actualParams['foo'] === 'bar';
-            });
-
-        $this->guzzleClient
-            ->get('http://www.example.com/foo', [
-                'query' => [
-                    'foo' => 'bar',
-                    'faz' => 'shazaam'
-                ]
-            ]);
-
-        $this->assertTrue($this->httpMock->verify());
-    }
-
-    /** @test */
-    public function queryParams_customLogic_notMatch_shouldFail()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->withQueryParams(function ($actualParams) {
-                return $actualParams['foo'] === 'notTheActualValueOfFoo';
-            });
-
-        $this->expectException(UnexpectedHttpRequestException::class);
-        $this->guzzleClient
-            ->get('http://www.example.com/foo', [
-                'query' => [
-                    'foo' => 'bar',
-                    'faz' => 'shazaam'
-                ]
-            ]);
-
-
-        $this->httpMock->verify();
-    }
-
-    /** @test */
-    public function bodyParams_notMatch_shouldFail()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->withBodyParams(['foo' => 'bar']);
-
-        $this->expectException(UnexpectedHttpRequestException::class);
-        $this->guzzleClient
-            ->get('http://www.example.com/foo', [
-                'form_params' => ['not' => 'what I expected']
-            ]);
-
-        $this->httpMock->verify();
-    }
-
-    /** @test */
-    public function bodyParams_match_shouldPass()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->withBodyParams(['foo' => 'bar']);
-
-        $this->guzzleClient
-            ->get('http://www.example.com/foo', [
-                'form_params' => ['foo' => 'bar']
-            ]);
-
-        $this->assertTrue($this->httpMock->verify());
-    }
-
-    /** @test */
-    public function body_customLogic_match_shouldPass()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->withBodyParams(function ($bodyParams) {
-                return $bodyParams['foo'] === 'bar';
-            });
-
-        $this->guzzleClient
-            ->get('http://www.example.com/foo', [
-                'form_params' => ['foo' => 'bar']
-            ]);
-
-        $this->assertTrue($this->httpMock->verify());
-    }
-
-    /** @test */
-    public function body_customLogic_notMatch_shouldFail()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->withBodyParams(function ($bodyParams) {
-                return $bodyParams['foo'] === 'bar';
-            });
-
-        $this->expectException(UnexpectedHttpRequestException::class);
-        $this->guzzleClient
-            ->get('http://www.example.com/foo', [
-                'form_params' => ['foo' => 'shablooey']
-            ]);
-
-        $this->httpMock->verify();
-    }
-
-    public function body_anyParams_shouldPass()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->withBodyParams(new Expect\Any());
-
-        $this->guzzleClient
-            ->get('http://www.example.com/foo', [
-                'form_params' => ['foo' => 'bar']
-            ]);
-
-        $this->httpMock->verify();
-    }
-
-    /** @test */
-    public function body_arrayContains_match_shouldPass()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->withBodyParams(new Expect\ArrayContains(['foo' => 'bar']));
-
-        $this->guzzleClient
-            ->get('http://www.example.com/foo', [
-                'form_params' => ['foo' => 'bar', 'faz' => 'baz']
-            ]);
-
-        $this->assertTrue($this->httpMock->verify());
-    }
-
-    /** @test */
-    public function body_arrayContains_notMatch_shouldFail()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->withBodyParams(new Expect\ArrayContains(['foo' => 'bar']));
-
-        $this->expectException(UnexpectedHttpRequestException::class);
-        $this->guzzleClient
-            ->get('http://www.example.com/foo', [
-                'form_params' => ['foo' => 'shablooey', 'faz' => 'baz']
-            ]);
-
-        $this->httpMock->verify();
-    }
-
-    /** @test */
-    public function body_outOfOrder_match_shouldPass()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->withBodyParams([
-                'faz' => 'baz',
-                'foo' => 'bar',
-            ]);
-
-        $this->guzzleClient
-            ->get('http://www.example.com/foo', [
-                'form_params' => [
-                    'foo' => 'bar',
-                    'faz' => 'baz',
-                ]
-            ]);
-
-        $this->assertTrue($this->httpMock->verify());
-    }
-
-    /** @test */
-    public function body_outOfOrder_nullValues_match_shouldPass()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->withBodyParams([
-                'faz' => 'baz',
-                'foo' => 'bar',
-                'nullA' => null,
-                'nullB' => null,
-            ]);
-
-        $this->guzzleClient
-            ->get('http://www.example.com/foo', [
-                'json' => [
-                    'faz' => 'baz',
-                    'foo' => 'bar',
-                    'nullB' => null,
-                    'nullA' => null,
-                ]
-            ]);
-
-        $this->assertTrue($this->httpMock->verify());
-    }
-
-    /** @test */
-    public function jsonBody_notMatch_shouldFail()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->withJsonBodyParams(['foo' => 'bar']);
-
-        $this->expectException(UnexpectedHttpRequestException::class);
-        $this->guzzleClient
-            ->get('http://www.example.com/foo', [
-                'json' => ['not' => 'what I expected']
-            ]);
-
-        $this->httpMock->verify();
-    }
-
-
-    /** @test */
     public function shouldFailIfNoRequestIsConfigured()
     {
         $this->guzzleClient
@@ -738,152 +314,6 @@ class MockTest extends TestCase
         $this->httpMock->verify();
     }
 
-    /** @test */
-    public function shouldFailIfTheRequestIsNotMade()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest();
-
-        $this->expectException(UnexpectedHttpRequestException::class);
-        try {
-            throw new Exception('too bad.');
-        } catch (Exception $ex) {
-            $this->httpMock->verify();
-            throw $ex;
-        }
-    }
-
-    /** @test */
-    public function shouldFailIfTheRequestIsMadeMoreThanOnce()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo');
-
-        $this->guzzleClient
-            ->get('http://www.example.com/foo');
-
-        $this->expectException(UnexpectedHttpRequestException::class);
-        $this->guzzleClient
-            ->get('http://www.example.com/foo');
-
-        $this->httpMock->verify();
-    }
-
-    /** @test */
-    public function shouldFailIfTheRequestIsMadeMoreThanTheSetTimes()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->times(2);
-
-        $this->guzzleClient
-            ->get('http://www.example.com/foo');
-        $this->guzzleClient
-            ->get('http://www.example.com/foo');
-
-        $this->expectException(UnexpectedHttpRequestException::class);
-        $this->guzzleClient
-            ->get('http://www.example.com/foo');
-
-        $this->httpMock->verify();
-    }
-
-    /** @test */
-    public function shouldFailIfTheRequestIsMadeLessThanTheSetTimes()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->times(3);
-
-        $this->guzzleClient
-            ->get('http://www.example.com/foo');
-        $this->guzzleClient
-            ->get('http://www.example.com/foo');
-
-        $this->expectException(UnexpectedHttpRequestException::class);
-        $this->httpMock->verify();
-    }
-
-    /** @test */
-    public function shouldPassIfTheRequestIsMadeAsManyTimesAsExpected()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->times(3);
-
-        $this->guzzleClient
-            ->get('http://www.example.com/foo');
-        $this->guzzleClient
-            ->get('http://www.example.com/foo');
-        $this->guzzleClient
-            ->get('http://www.example.com/foo');
-
-        $this->assertTrue($this->httpMock->verify());
-    }
-
-    /** @test */
-    public function shouldPassIfTheRequestIsMadeOnce()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo');
-
-        $this->guzzleClient
-            ->get('http://www.example.com/foo');
-
-        $this->assertTrue($this->httpMock->verify());
-    }
-
-    /**
-     * @testWith [0,1,2,5]
-     * @param int $nrOfRequests
-     * @throws CompoundUnexpectedHttpRequestException
-     */
-    public function shouldPassRegardlessOfTheNumberOfRequests(int $nrOfRequests)
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->zeroOrMoreTimes();
-
-        /** @noinspection PhpUnusedLocalVariableInspection */
-        foreach (range(0, $nrOfRequests) as $i) {
-            $this->guzzleClient
-                ->get('http://www.example.com/foo');
-        }
-
-        $this->assertTrue($this->httpMock->verify());
-    }
-
-
-    /** @test */
-    public function shouldPassIfTheRequestIsMadeTheNumberOfSetTimes()
-    {
-        $this->httpMock
-            ->shouldReceiveRequest()
-            ->withMethod('GET')
-            ->withUrl('http://www.example.com/foo')
-            ->times(3);
-
-        $this->guzzleClient
-            ->get('http://www.example.com/foo');
-        $this->guzzleClient
-            ->get('http://www.example.com/foo');
-        $this->guzzleClient
-            ->get('http://www.example.com/foo');
-
-        $this->assertTrue($this->httpMock->verify());
-    }
 
     /** @test */
     public function shouldResponseUsingClosure()
